hugging-face-agent/requirements.txt
# requirements.txt (for Hugging Face Space)
# Core dependencies
smolagents>=0.3.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
gradio>=4.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-magic>=0.4.27

# NLP libraries
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Utilities
pydantic>=2.0.0
loguru>=0.7.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# Optional for better performance
yake>=0.4.8

smolagents>=0.3.0
litellm>=1.0.0
gradio>=4.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
scipy>=1.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
